+++ packages/check-types-mini/changelog.md
@@ -4,13 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.3.0] - 2017-05-22
+### Added
+- ✨ `opts.acceptArrays` will accept arrays too, if they contain only the same type elements as the one that's being checked.
+- ✨ `opts.acceptArraysIgnore` - lets you ignore per-key level when `opts.acceptArrays` is on. 👍
+
 ## [1.2.0] - 2017-05-15
 ### Added
-- opts.ignoreKeys won't throw now if input is a single string.
+- `opts.ignoreKeys` won't throw now if input is a single string.
 
 ## [1.1.0] - 2017-05-15
 ### Added
-- opts.ignoreKeys
+- ✨ `opts.ignoreKeys`
 
 ## 1.0.0 - 2017-05-15
 ### New
@@ -18,3 +23,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 [1.1.0]: https://github.com/codsen/check-types-mini/compare/v1.0.1...v1.1.0
 [1.2.0]: https://github.com/codsen/check-types-mini/compare/v1.1.0...v1.2.0
+[1.3.0]: https://github.com/codsen/check-types-mini/compare/v1.2.2...v1.3.0